fixed issue with wrong type being passed to method

+++ DMCircularScrollView/DMCircularScrollView/DMCircularScrollView.m
@@ -196,7 +196,7 @@
     }
 }
 
-- (void)scrollViewDidEndZooming:(UIScrollView *)sv withView:(UIView *)view atScale:(float)scale
+- (void)scrollViewDidEndZooming:(UIScrollView *)sv withView:(UIView *)view atScale:(double)scale
 {
     if (self.scrollViewDelegate && [self.scrollViewDelegate respondsToSelector:@selector(scrollViewDidEndZooming:withView:atScale:)])
     {
